requirements.txt
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
fpdf==1.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
paramiko==3.5.1
port-scanner==0.1.4
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
PyNaCl==1.5.0
requests==2.32.3
rich==13.9.4
scapy==2.6.1
setuptools==76.0.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
vuln_scanner==1.0.2
Werkzeug==3.1.3
wheel==0.45.1
